Inline the tier requirements
